hw03_without_point_6_optional

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepository.java
@@ -16,7 +16,7 @@ import ru.javawebinar.topjava.repository.MealRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Repository
+@Repository("JdbcMealRepository")
 public class JdbcMealRepository implements MealRepository {
     private static final BeanPropertyRowMapper<Meal> ROW_MAPPER = BeanPropertyRowMapper.newInstance(Meal.class);
 
@@ -48,8 +48,8 @@ public class JdbcMealRepository implements MealRepository {
             Number newKey = insertMeal.executeAndReturnKey(map);
             meal.setId(newKey.intValue());
         } else if (namedParameterJdbcTemplate.update(
-                "UPDATE meals SET date_time=:dateTime, description=:description, calories=:calories, " +
-                        "user_id=:userId WHERE id=:id", map) == 0) {
+                "UPDATE meals SET date_time=:dateTime, description=:description, calories=:calories " +
+                        "WHERE id=:id AND user_id=:userId", map) == 0) {
             return null;
         }
         return meal;
@@ -57,13 +57,13 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id) != 0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=?", id, userId) != 0;
     }
 
     @Override
     public Meal get(int id, int userId) {
         List<Meal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=? AND user_id=?", ROW_MAPPER, id, userId);
-        return DataAccessUtils.requiredSingleResult(meals);
+        return DataAccessUtils.uniqueResult(meals);
     }
 
     @Override
@@ -73,9 +73,8 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public List<Meal> getBetweenHalfOpen(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId) {
-        List<Meal> meals = jdbcTemplate.query(
-                "SELECT * FROM meals WHERE user_id=? AND date_time>=? AND date_time<?",
+        return jdbcTemplate.query(
+                "SELECT * FROM meals WHERE user_id=? AND date_time>=? AND date_time<? ORDER BY id",
                 ROW_MAPPER, userId, startDateTime, endDateTime);
-        return meals;
     }
 }
